sched/fair: Consider only idle CPUs for active migration
A task that does not fit on its current CPU is actively migrated
to a higher capacity CPU from the scheduler tick path. Since
active migration involves stopping the currently running task
and migrating to a different CPU, it is better to do this only
if the target CPU is idle. The task does not have to necessarily
run on the lower capacity cluster until the next tick. When a
higher capacity CPU becomes idle, it can pull the misfit task
running on the lower capacity CPU.

Change-Id: Idda0bd6ac8cc4bc22f5eddc69236014d04708ecd
Signed-off-by: Pavankumar Kondeti <pkondeti@codeaurora.org>