sched: Fix a compilation issue for !SCHED_WALT
A CPU is marked as reserved when a task is under active migration to
that CPU. The reserved CPUs are excluded from CPU selection in
task placement. This reservation scheme is currently tied to WALT.
So add a stub function for is_reserved() to avoid compilation issue
when WALT is disabled.

Change-Id: I77c6a8fd3f24f1763fd97fce18574613e9e947ff
Signed-off-by: Pavankumar Kondeti <pkondeti@codeaurora.org>