coresight: Fix the mem_type and  mem_size of ETR
The ETR sysfs nodes mem_type and mem_size can be altered
when etr is in enabled state. But that does not the change
the operation of ETR, as its already initialised with
settings at the point of enablement.This fix check the ETR
state, when user tries to change them on runtime.

Signed-off-by: Saranya Chidura <schidura@codeaurora.org>
Change-Id: Iba8c695dfa00d6183e88aab8316f16749d8da875