power: smb5: add support to enable/disable hvdcp optimization
On certain PMICs, Flash has a requirement to limit VBUS at 5V
when it is active. This is to prevent the VIN_FLASH from seeing
a high input voltage. Add support for this by forcing VBUS to 5V
when there is a notification from flash driver based on flash_active
power-supply property. The userspace is also notified to stop any
further VBUS optimizations by setting the hvdcp_opti_allowed property.
The optimization is re-enabled when flash_active is cleared by the flash
driver.

Change-Id: I3748fe75f467969ecc4ec2a60c5360dba13953f1
Signed-off-by: Ashay Jaiswal <ashayj@codeaurora.org>