<DONOT MERGE>kthread/smpboot: Serialize kthread parking against wakeup
> On 5/1/2018 5:01 PM, Peter Zijlstra wrote:

> > Let me ponder what the best solution is, it's a bit of a mess.

So there's:

 - TASK_PARKED, which we can make a special state; this solves the
   problem because then wait_task_inactive() is guaranteed to see
   TASK_PARKED and DTRT.

 - complete(&kthread->parked), which we can do inside schedule(); this
   solves the problem because then kthread_park() will not return early
   and the task really is blocked.

and I'm fairly sure I thought of a 3rd way to cure things, but now that
I'm writing things down I cannot seem to remember :/ -- could be we muck
with wait_task_inactive().

In any case, I hate all of them, but I think the completion one is the
least horrible because it gives the strongest guarantees and cleans up
most. But promoting TASK_PARKED to special certainly is the earier
patch.

The below boots, but that's about all I did with it. Opinions?

Change-Id: Ie71645d37046f7ee74df880dbead29efbaad199a
Signed-off-by: Gaurav Kohli <gkohli@codeaurora.org>