sched/fair: Expand the search in energy_aware_wake_cpu()
energy_aware_wake_cpu() selects a sched group based on the task
utilization and its preferred cluste. The search for the best
CPU is limited only to this primary sched group on a big little
system. However this results in a suboptimal performance when
all the CPUs in the primary sched group are overutilized. Hence
expand the search to the other sched groups in this case.

Change-Id: If914b9af5644194f41439262898ebaced6aedc33
Signed-off-by: Pavankumar Kondeti <pkondeti@codeaurora.org>