usb: phy-msm-qusb-v2: Override TUNE1 while honoring efuse value
We need to set BIT:7 of TUNE1 register and DEBUG_CTRL1 to 0x09
for host mode chirp fix while honoring the efuse value. Now we
are checking if efuse value is present or not, based on that we
are programming TUNE1 value. This change also adds a property
on which we can decide to override BIAS_CTRL2 from driver.
This change also adds support to override tuneX paramerters
through debugfs in HOST mode.

Change-Id: I2f562391464b8a035b6dec23d1b909765cb4613a
Signed-off-by: Pratham Pratap <prathampratap@codeaurora.org>