power: smb5-lib: load HW JEITA config in general psy events
Currently, JEITA configuration is updated with a BMS event. There
is a chance that BMS is not updating with a long time and that would
cause the JEITA configuration is not loaded and JEITA is mis-triggered,
so update the JEITA configuration in general power supply events and
defer the updating if BMS power supply is not ready.

Change-Id: Ib3b792ea8fc64083ea14b7ea5a00565815a0cec6
Signed-off-by: Fenglin Wu <fenglinw@codeaurora.org>