cpuidle: lpm-levels: Hold lock during tick_broadcast_enter
When a core is doing system pm enter, as a part of
the GIC workaround, it would save the distributor registers.
There is a window where other core is doing tick_broadcast_enter
which will disable broadcast timer interrupt and enable back.
If the GIC save for broadcast timer interrupt falls in this window
it would be saved as disabled which leads to GIC restore would
keep interrupt disabled after exiting from low power mode.

To avoid this, hold a lock during tick_broadcast_enter and
system pm enter.

Change-Id: I924aa1a14f8fe4dc55a0e0d90e82e7c0ace6ec4e
Signed-off-by: Srinivas Rao L <lsrao@codeaurora.org>
Signed-off-by: Maulik Shah <mkshah@codeaurora.org>