watchdog: Add qcom VM watchdog driver
Watchdog timer is configured with bark and bite time interval.
If VM fails to reset the counter during bark interval, a bark
interrupt (SPI) is send to the VM. Failing to respond to the
bark interrupt causes a bite, when counter reaches the bite
interval. The bite interrupt is routed to secure world, and
causes a system reset.

Change-Id: Ia71a5b1f5e18382fd3c7fb00818c80c150ae1ba5
Signed-off-by: Neeraj Upadhyay <neeraju@codeaurora.org>