binder: create node flag to request sender's security context
To allow servers to verify client identity, allow a node
flag to be set that causes the sender's security context
to be delivered with the transaction. The BR_TRANSACTION
command is extended in BR_TRANSACTION_SEC_CTX to
contain a pointer to the security context string.

Bug: 25646100
Change-Id: I0ec053072d6337576680067e07f90dd054b1ecfb
Signed-off-by: Todd Kjos <tkjos@google.com>
Git-repo: https://android.googlesource.com/kernel/msm
Git-commit: c0b9c16ef74517f5ff5e9404fa38b7f51269c00e
Signed-off-by: Dennis Cagle <dcagle@codeaurora.org>