drm/msm: check HDMI HFVSDB block before adding formats
Currently, the EDID parser adds the formats based on the
parsing of the Video data blocks and other CTA blocks.

However, there is no input validation based on the
HDMI HFVSDB block to check whether the mode advertised
by the sink actually falls in the TMDS char rate limits.

Add this check in the EDID parser to make sure invalid
formats are not added to the list.

Change-Id: I092d9e113a99fae4983ced7fa77472d1eeef8575
Signed-off-by: Abhinav Kumar <abhinavk@codeaurora.org>
Signed-off-by: Narender Ankam <nankam@codeaurora.org>
Signed-off-by: Anusha Shakarad <ashaka@codeaurora.org>