xhci: Prevent invalid access of virt_dev after free up
The xhci_free_virt_device sets the xhci->devs[*] to
NULL as part of tearing down the host stack.
If tearing down the host stack races with the transfers
with the transfers over a storage device, it can lead
to a NULL pointer derefrence as part of checking the
ep_state.
Fix this by checking the validity of the xhci->devs
under the spinlock before queuing on the endpoint.

Change-Id: I166b54af93c6e1844bff6244fb790c2a16b70203
Signed-off-by: Sriharsha Allenki <sallenki@codeaurora.org>