drivers: qcom: Add support for subsystem backup and restore
Remote subsystem can backup-restore data with the help of HLOS.
For this requirement, add a helper driver in the kernel.

This driver implements the QMI client for interaction with the
remote subsystem. Upon a backup/restore notification, this driver
allocates a shared buffer and locks exclusively for the subsystem
to use. Upon completion of backup/restore, the shared buffer is
freed.

Userspace can read/write the backup image using /dev interface
exposed by this driver.

Change-Id: I137d2ca331f39af3ac9d5694a2660c637de35c5a
Signed-off-by: Shreyas K K <shrekk@codeaurora.org>