usb: dwc3-msm: Reset trb_dequeue to 0 in gsi EP configuration
On certain platforms, there are GSI usecase EPs which use the SW
path, i.e., they trigger completion IRQs to the CPU and not the
USB-GSI wrapper. Consider a case where such a device is connected
to a host and 5 ping packets have been received from the host.
The dequeue pointer index is at the 5th TRB. Now the device gets
re-enumerated on the host due to some reason. The dequeue index
is still at 5 but the transfers from host will start afresh from
the 0th TRB. This will lead to the device replying with pong to
the 1st ping packet after 5 seconds which is unacceptable.
Fix this by resetting trb_dequeue index to 0 in gsi_configure_ep.

Change-Id: I0665cda2a1558b0505b5c4aecc47e738ece16789
Signed-off-by: Ajay Agarwal <ajaya@codeaurora.org>