net: ipc_router: Make wakeup timeout configurable
The ipc router driver can guarantee that the packet gets queued to the
socket but cannot guarantee the client process will get time to run if
auto sleep is enabled. This config will help mitigate missed packets on
systems where auto sleep is too aggressive.

Use the __pm_wakeup_event() api so a hard wakeup can be specified.
This will ensure a resume occurs if the data coming in happens while
the device is going into suspend.

Change-Id: Ifac404b5ef4561574eb0e6a5c4bb20f5e741b483
Signed-off-by: Arun Prakash <app@codeaurora.org>