msm: camera: reqmgr: check if link handle is correctly passed
Instead of the link handle if the dev handle is
passed for dumping the request information, this
can lead to accessing invalid data structure.
To avodi accessing invalid data structure based on
the dev handle, first check if the link handle passed
in IOCTL is matchting with looked up link handle.

CRs-Fixed: 3097336
Change-Id: I59bb529c7caf82e6046844640ba6600fc9c21a0a
Signed-off-by: Tejas Prajapati <quic_tpraja@quicinc.com>