usb: dwc3: Add support for RT threaded interrupt for dwc3
In use cases where HID interfaces are being used to send data
periodically, dwc3 hard irq (the actual interrupt from the GIC)
is coming very early but the actual retiring of the request is
getting delayed due to worker thread scheduling. This latency
is causing data drop to occur. Adding support for Threaded IRQ
to be used for such requirements on specific targets based on
device tree.

Change-Id: Ifb76046096f2647e39e07454168d470858838c3f
Signed-off-by: Krishna Kurapati <quic_kriskura@quicinc.com>